solve_user_borrower_guarantor_attachments

+++ database/migrations/2025_03_09_112831_create_borrowers_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('borrowers', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->integer('phone');
-            $table->integer('nationalID');
+            $table->bigInteger('phone');
+            $table->bigInteger('nationalID');
             $table->string('address');
             $table->string('job');
             $table->timestamps();
